REFACTOR: directory rule 변경
