+++ joonggo-app/src/app/products/[[...category]]/page.tsx
@@ -1,11 +1,10 @@
-import CategoryLevel from "@/app/products/[[...category]]/category-level";
-import ProductList from "@/app/products/[[...category]]/product-list";
+import CategoryLevel from "@/app/products/[[...category]]/_components/category-level";
+import ProductList from "@/app/products/[[...category]]/_components/product-list";
 
 export const dynamicParams = false; // Dynamic segments not included in generateStaticParams will return a 404.
 
 export async function generateStaticParams() {
     return [{category: ["elect", "TV"]}, {category: ["elect", "TV", "samsung"]}, {category: ["elect"]}, {category: [""]}];
-
 }
 
 export default function Page(params: { params: { category: string[] } }) {
