+++ joonggo-app/src/app/types/Product.ts
@@ -1,0 +1,5 @@
+export interface Product {
+    id: string,
+    name: string,
+    price: number
+}
