+++ joonggo-app/src/app/types/SearchParam.ts
@@ -1,0 +1,7 @@
+export class SearchParam {
+    searchValue: string;
+
+    constructor(searchValue: string) {
+        this.searchValue = searchValue;
+    }
+}
